+++ core/b_plus_tree/b_plus_tree.hpp
@@ -148,9 +148,9 @@ namespace core::b_plus_tree {
 
         //template<typename T, typename Serializer>
         //bool append(T item, Serializer serializer);
-        bool append(data_ptr_t data, size_t size);
+        bool append(data_ptr_t data, uint32_t size);
         bool append(item_data item);
-        bool remove(data_ptr_t data, size_t size);
+        bool remove(data_ptr_t data, uint32_t size);
         bool remove(item_data item);
         //template<typename T>
         //bool remove_index(T value); // transforms value to index_t
